Revert "Uart improvments (#1024)"
This reverts commit 82db46db34db21025f0f0fb2601766db6dd38134.
