+++ esphome/components/uart/__init__.py
@@ -28,25 +28,13 @@ def validate_rx_pin(value):
     return value
 
 
-UARTParityOptions = uart_ns.enum('UARTParityOptions')
-UART_PARITY_OPTIONS = {
-    'NONE': UARTParityOptions.UART_CONFIG_PARITY_NONE,
-    'EVEN': UARTParityOptions.UART_CONFIG_PARITY_EVEN,
-    'ODD': UARTParityOptions.UART_CONFIG_PARITY_ODD,
-}
-
 CONF_STOP_BITS = 'stop_bits'
-CONF_DATA_BITS = 'data_bits'
-CONF_PARITY = 'parity'
-
 CONFIG_SCHEMA = cv.All(cv.Schema({
     cv.GenerateID(): cv.declare_id(UARTComponent),
     cv.Required(CONF_BAUD_RATE): cv.int_range(min=1),
     cv.Optional(CONF_TX_PIN): pins.output_pin,
     cv.Optional(CONF_RX_PIN): validate_rx_pin,
     cv.Optional(CONF_STOP_BITS, default=1): cv.one_of(1, 2, int=True),
-    cv.Optional(CONF_DATA_BITS, default=8): cv.int_range(min=5, max=8),
-    cv.Optional(CONF_PARITY, default="NONE"): cv.enum(UART_PARITY_OPTIONS, upper=True)
 }).extend(cv.COMPONENT_SCHEMA), cv.has_at_least_one_key(CONF_TX_PIN, CONF_RX_PIN))
 
 
@@ -62,8 +50,6 @@ def to_code(config):
     if CONF_RX_PIN in config:
         cg.add(var.set_rx_pin(config[CONF_RX_PIN]))
     cg.add(var.set_stop_bits(config[CONF_STOP_BITS]))
-    cg.add(var.set_data_bits(config[CONF_DATA_BITS]))
-    cg.add(var.set_parity(config[CONF_PARITY]))
 
 
 # A schema to use for all UART devices, all UART integrations must extend this!
